+++ workers/tests/integrationtests/test_api_regions.py
@@ -21,6 +21,12 @@ class TestApiIndicator(unittest.TestCase):
         self.assertEqual(response.status_code, 200)
         self.assertTrue(geojson.loads(response.content).is_valid)
 
+    def test_get_available_layer_indicator_combinations(self):
+        url = "/indicatorLayerCombinations"
+        response = self.client.get(url)
+        self.assertEqual(response.status_code, 200)
+        self.assertIsInstance(geojson.loads(response.content), dict)
+
 
 if __name__ == "__main__":
     unittest.main()